fix components
